requirements.txt
fastapi
uvicorn
pptx
geminiai
google.generativeai